Rename createPortal.js to portals.js
